fix: prevent OSError: read-only file system
I found that there is a chance of OSError to occur when we try to load models into a temporary directory such as in the strictly managed environment such as managed containers on the cloud.

This commit will prevent that to happen. The code for the new function 'download' is originated from wget library written by anatoly techtonik with slight revision done by me.